% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExcel.R
\name{makeExcel}
\alias{makeExcel}
\title{Output a set of data tables to an accessible xlsx file}
\usage{
makeExcel(tables, titles, subtitles, tabnames, tablenames, filename)
}
\arguments{
\item{tables}{A list of data frames. Each list element should be a data.frame.}

\item{titles}{A list of table titles. Each list element should be a character vector of length 1.}

\item{subtitles}{A list of subtitles. Each list element should be a character vector of any length.}

\item{tabnames}{A list of worksheet tab names. Each list element should be a character vector of length 1.}

\item{tablenames}{A list of table names. Each list element should be a character vector of length 1.}

\item{filename}{A string specifying the name of the output file.}
}
\value{
A .xlsx workbook
}
\description{
This function outputs a .xlsx file made up of a series of R data frames.
}
